Move connector update logic into connector manager
Clean up connector ids update hack in main.
Add ConnectorManager::updateConnectorIds.
